Moving HttpClient interceptors from the OpenTracing service to commons utilities
