+++ repose-aggregator/commons/commons-utilities/build.gradle
@@ -13,6 +13,7 @@ dependencies {
     compile "org.jvnet.jaxb2_commons:jaxb2-basics-runtime"
     compile "com.typesafe.play:play-json_$scalaMajDotMin"
     compile "io.opentracing:opentracing-api"
+    compile "com.uber.jaeger:jaeger-apachehttpclient"
 
     compileOnly "javax.ws.rs:javax.ws.rs-api"
     compileOnly "org.glassfish:javax.servlet" //todo: swap these as part of REP-3950
