+++ repose-aggregator/components/services/http-client-service/http-client-service-impl/build.gradle
@@ -2,10 +2,10 @@ dependencies {
     compile project(":repose-aggregator:core:repose-core-api")
     compile project(":repose-aggregator:components:services:http-client-service:http-client-service-api")
     compile project(":repose-aggregator:components:services:health-check-service:health-check-service-api")
-    compile project(":repose-aggregator:components:services:open-tracing-service:open-tracing-service-api")
     compile "org.slf4j:slf4j-api"
     compile "org.slf4j:jcl-over-slf4j"
     compile "org.springframework:spring-beans"
+    compile "io.opentracing:opentracing-api"
 
     compileOnly "javax.inject:javax.inject"
