+++ repose-aggregator/components/services/http-client-service/http-client-service-impl/src/main/java/org/openrepose/core/services/httpclient/impl/HttpConnectionPoolProvider.java
@@ -19,6 +19,8 @@
  */
 package org.openrepose.core.services.httpclient.impl;
 
+import com.uber.jaeger.httpclient.TracingResponseInterceptor;
+import io.opentracing.Tracer;
 import org.apache.http.Header;
 import org.apache.http.client.HttpClient;
 import org.apache.http.client.config.CookieSpecs;
@@ -37,9 +39,9 @@ import org.apache.http.params.CoreConnectionPNames;
 import org.apache.http.params.HttpParams;
 import org.apache.http.ssl.SSLContextBuilder;
 import org.apache.http.ssl.SSLContexts;
+import org.openrepose.commons.utils.opentracing.httpclient.ReposeTracingRequestInterceptor;
 import org.openrepose.core.service.httpclient.config.HeaderType;
 import org.openrepose.core.service.httpclient.config.PoolType;
-import org.openrepose.core.services.opentracing.OpenTracingService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -62,7 +64,7 @@ public final class HttpConnectionPoolProvider {
     private HttpConnectionPoolProvider() {
     }
 
-    public static HttpClient genClient(String configRoot, PoolType poolConf, OpenTracingService openTracingService) {
+    public static HttpClient genClient(String configRoot, PoolType poolConf, Tracer tracer) {
         PoolingClientConnectionManager cm = new PoolingClientConnectionManager();
 
         cm.setDefaultMaxPerRoute(poolConf.getHttpConnManagerMaxPerRoute());
@@ -90,11 +92,11 @@ public final class HttpConnectionPoolProvider {
         //Pass in the params and the connection manager
         DefaultHttpClient client = new DefaultHttpClient(cm, params);
 
-        // OpenTracing capabilities.  Only add http interceptors when opentracing service is enabled
-        if (openTracingService.isEnabled()) {
-            client.addRequestInterceptor(openTracingService.getRequestInterceptor());
-            client.addResponseInterceptor(openTracingService.getResponseInterceptor());
-        }
+        // OpenTracing support
+        // Note that although we always register these interceptors, the provided Tracer may be a NoopTracer,
+        // making nearly all of the work done by these interceptors a no-op.
+        client.addRequestInterceptor(new ReposeTracingRequestInterceptor(tracer));
+        client.addResponseInterceptor(new TracingResponseInterceptor());
 
         SSLContext sslContext = ProxyUtilities.getTrustingSslContext();
